Rename `classes` to `core`
