--- docs/source/api/classes.rst
@@ -1,16 +1,0 @@
-###############
-classes package
-###############
-
-.. rubric:: Modules
-
-.. autosummary::
-   :toctree: classes
-
-   ~xgi.classes.hypergraph
-   ~xgi.classes.dihypergraph
-   ~xgi.classes.simplicialcomplex
-   ~xgi.classes.reportviews
-   ~xgi.classes.direportviews
-   ~xgi.classes.hypergraphviews
-   ~xgi.classes.function
