--- docs/source/api/classes/xgi.classes.dihypergraph.rst
@@ -1,14 +1,0 @@
-﻿xgi.classes.dihypergraph
-========================
-
-.. currentmodule:: xgi.classes.dihypergraph
-
-.. automodule:: xgi.classes.dihypergraph
-
-   .. rubric:: Classes
-
-   .. autosummary::
-      :toctree: .
-      :nosignatures:
-      
-        DiHypergraph
