--- docs/source/api/classes/xgi.classes.direportviews.rst
@@ -1,16 +1,0 @@
-﻿xgi.classes.direportviews
-=========================
-
-.. currentmodule:: xgi.classes.direportviews
-
-.. automodule:: xgi.classes.direportviews
-
-   .. rubric:: Classes
-
-   .. autosummary::
-      :toctree: .
-      :nosignatures:
-
-        DiIDView
-        DiNodeView
-        DiEdgeView
