--- docs/source/api/classes/xgi.classes.hypergraph.rst
@@ -1,14 +1,0 @@
-﻿xgi.classes.hypergraph
-======================
-
-.. currentmodule:: xgi.classes.hypergraph
-
-.. automodule:: xgi.classes.hypergraph
-
-   .. rubric:: Classes
-
-   .. autosummary::
-      :toctree: .
-      :nosignatures:
-      
-        Hypergraph
