--- docs/source/api/classes/xgi.classes.hypergraphviews.rst
@@ -1,10 +1,0 @@
-﻿xgi.classes.hypergraphviews
-===========================
-
-.. currentmodule:: xgi.classes.hypergraphviews
-
-.. automodule:: xgi.classes.hypergraphviews
-
-   .. rubric:: Functions
-
-   .. autofunction:: subhypergraph
