--- docs/source/api/classes/xgi.classes.simplicialcomplex.rst
@@ -1,14 +1,0 @@
-﻿xgi.classes.simplicialcomplex
-=============================
-
-.. currentmodule:: xgi.classes.simplicialcomplex
-
-.. automodule:: xgi.classes.simplicialcomplex
-
-   .. rubric:: Classes
-
-   .. autosummary::
-      :toctree: .
-      :nosignatures:
-      
-        SimplicialComplex
