+++ docs/source/api/core.rst
@@ -1,0 +1,16 @@
+##################
+core functionality
+##################
+
+.. rubric:: Modules
+
+.. autosummary::
+   :toctree: core
+
+   ~xgi.core.hypergraph
+   ~xgi.core.dihypergraph
+   ~xgi.core.simplicialcomplex
+   ~xgi.core.reportviews
+   ~xgi.core.direportviews
+   ~xgi.core.hypergraphviews
+   ~xgi.core.function
