+++ docs/source/api/core/xgi.core.dihypergraph.rst
@@ -1,0 +1,14 @@
+﻿xgi.core.dihypergraph
+=====================
+
+.. currentmodule:: xgi.core.dihypergraph
+
+.. automodule:: xgi.core.dihypergraph
+
+   .. rubric:: Classes
+
+   .. autosummary::
+      :toctree: .
+      :nosignatures:
+      
+        DiHypergraph
