+++ docs/source/api/core/xgi.core.direportviews.rst
@@ -1,0 +1,16 @@
+﻿xgi.core.direportviews
+======================
+
+.. currentmodule:: xgi.core.direportviews
+
+.. automodule:: xgi.core.direportviews
+
+   .. rubric:: Classes
+
+   .. autosummary::
+      :toctree: .
+      :nosignatures:
+
+        DiIDView
+        DiNodeView
+        DiEdgeView
