+++ docs/source/api/core/xgi.core.hypergraph.rst
@@ -1,0 +1,14 @@
+﻿xgi.core.hypergraph
+===================
+
+.. currentmodule:: xgi.core.hypergraph
+
+.. automodule:: xgi.core.hypergraph
+
+   .. rubric:: Classes
+
+   .. autosummary::
+      :toctree: .
+      :nosignatures:
+      
+        Hypergraph
