+++ docs/source/api/core/xgi.core.hypergraphviews.rst
@@ -1,0 +1,10 @@
+﻿xgi.core.hypergraphviews
+========================
+
+.. currentmodule:: xgi.core.hypergraphviews
+
+.. automodule:: xgi.core.hypergraphviews
+
+   .. rubric:: Functions
+
+   .. autofunction:: subhypergraph
