+++ docs/source/api/core/xgi.core.simplicialcomplex.rst
@@ -1,0 +1,14 @@
+﻿xgi.core.simplicialcomplex
+==========================
+
+.. currentmodule:: xgi.core.simplicialcomplex
+
+.. automodule:: xgi.core.simplicialcomplex
+
+   .. rubric:: Classes
+
+   .. autosummary::
+      :toctree: .
+      :nosignatures:
+      
+        SimplicialComplex
