+++ xgi/core/__init__.py
