+++ xgi/core/hypergraphviews.py
